requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.10
langgraph>=0.0.10
python-multipart>=0.0.6
jinja2>=3.1.2
pydantic>=2.0.0
langgraph-checkpoint-sqlite
